Expiry trigger only when its scheduler is defunct.
For long running jobs it may happen that a trigger has passed its expiry
threshold, but still be valid - just a job is running long. In such case
we check if its owning scheduler is alive and if so we assume it's the
case. Only when scheduler is defunct we can treat the trigger's lock as
expired.