Make it possible to opt out of JVM shutdown on cluster check in failure
org.quartz.jobStore.checkInErrorHandler.class is a new property that can
be used to specify the error handler class. It's currently just a Runnable,
in order to property pause and unpause Quartz we likely will have to introduce
a more sophisticated implementation or at least pass in some parameters down
the road.

However, this makes it possible to opt out of the controversial behavior.

Closes #147.